requirements.txt
SQLAlchemy~=2.0.31
python-dotenv~=1.0.1
aiogram~=3.10.0